[bug #157977863]fix directory
