+++ .pytest_cache/v/cache/lastfailed
@@ -1,11 +1,8 @@
 {
-  "tests/test_api-tracker.py": true,
   "tests/test_api-tracker.py::MaintenanceTrackerApiTest::test_create_request": true,
   "tests/test_api-tracker.py::MaintenanceTrackerApiTest::test_delete_request": true,
   "tests/test_api-tracker.py::MaintenanceTrackerApiTest::test_read_empty_request": true,
   "tests/test_api-tracker.py::MaintenanceTrackerApiTest::test_read_request": true,
   "tests/test_api-tracker.py::MaintenanceTrackerApiTest::test_read_request_with_id": true,
-  "tests/test_api-tracker.py::MaintenanceTrackerApiTest::test_update_requests": true,
-  "tests/test_api-tracker.py::MaintenanceTrackerApiTest::test_user_signin": true,
-  "tests/test_api-tracker.py::MaintenanceTrackerApiTest::test_user_signup": true
+  "tests/test_api-tracker.py::MaintenanceTrackerApiTest::test_update_requests": true
 }
