+++ run.py
@@ -1,7 +1,9 @@
 #run.py
 
 
-from app import app
+from app import create_app
+
+app = create_app('development')
 
 if __name__ == '__main__':
     app.run()
